refactor: move out element constructors to separate module
